% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{generate_token}
\alias{generate_token}
\title{Generate User Token}
\usage{
generate_token()
}
\value{
A character string representing the unique token.
}
\description{
This function generates a unique token based on the hard drive serial number.
The token can be sent to the package maintainer for authorization.
}
