% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{get_hard_drive_serial}
\alias{get_hard_drive_serial}
\title{Get Hard Drive Serial Number}
\usage{
get_hard_drive_serial()
}
\value{
A character string of the hard drive serial number.
}
\description{
This function retrieves the hard drive serial number.
It works on Windows, Linux, and macOS systems.
}
