% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_missense_snp.R
\name{get_missense_snp}
\alias{get_missense_snp}
\title{查询基因的错义突变}
\usage{
get_missense_snp(gene_name)
}
\arguments{
\item{gene_name}{基因名称（如："ACTN3"）}
}
\value{
返回该基因的错义突变信息，并将其保存为 CSV 文件
}
\description{
查询基因的错义突变
}
