% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FASTchemistry.R
\name{install_chemistry}
\alias{install_chemistry}
\title{Install Chemistry Package}
\usage{
install_chemistry()
}
\description{
This function installs the MR-SPI package from GitHub with user-friendly messages displaying "Chemistry".
}
