% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authorization.R
\name{validate_user}
\alias{validate_user}
\title{Validate User Authorization}
\usage{
validate_user()
}
\value{
TRUE if authorized, otherwise stops with an error.
}
\description{
This function checks if the user is authorized based on their token.
}
